+++ mitra_federation/src/http_client.rs
@@ -14,7 +14,7 @@ use mitra_utils::urls::{get_hostname, UrlError};
 use super::agent::FederationAgent;
 
 const CONNECTION_TIMEOUT: u64 = 30;
-const REDIRECT_LIMIT: usize = 3;
+pub const REDIRECT_LIMIT: usize = 3;
 
 // See also: mitra_validators::posts::CONTENT_MAX_SIZE
 pub const RESPONSE_SIZE_LIMIT: usize = 1_000_000;