+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed parsing of emoji shortcodes.
+- Re-sign fetcher request after redirection.
 
 ## [2.16.0] - 2024-04-16